Book: Move class declarations to the header and expose g_playerBook
